AI/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
langchain
langchain-community
langchain-openai
langchain-text-splitters
openai
pydantic
pgvector
psycopg2-binary